Rewrite the toolbar generation code
The old toolbar code didn't actually support any nontrivial
customization of toolbars, as it had a hardcoded list of commands which
would be enabled or disabled, and the checks for enabling/disabling
toolbar buttons was done only in the specific situations which those
commands used.

Make icon size a user customizable option and add support for
dynamically changing it at runtime.

git-svn-id: svn+ssh://aegisub/var/svn/aegisub/trunk@5468 93549f3f-7f0a-0410-a4b3-e966c9c94f04