Rewrite the font collector
The fontconfig collector should now always pick the exact font files
used by libass rather than a giant mishmash of vaguely related files
which may or may not include the correct font.

Make the freetype font collector windows-only, as it's far inferior to
the fontconfig collector and is only present as a fallback.

Add option to copy the fonts to the script's folder.

Closes #1059.

git-svn-id: svn+ssh://aegisub/var/svn/aegisub/trunk@6279 93549f3f-7f0a-0410-a4b3-e966c9c94f04