Move the visual tool submode buttons to the main visual tool toolbar
Creating an entire toolbar for one button in the drag mode is a
pointless waste of space and source of ugliness. The secondary toolbar
was more reasonable for the vector clip tool, but it's still a bit
pointless as there's enough space for it on the main toolbar at any
reasonable zoom level.

Closes #1401.

git-svn-id: svn+ssh://aegisub/var/svn/aegisub/trunk@6333 93549f3f-7f0a-0410-a4b3-e966c9c94f04