Add support for setting the active line from automation macros
The macro processing function can now return a second value, which is
the index of the line to make active, which must be one of the lines in
the selection. If it is not, or if the value is not present, then the
active line is left unchanged if it is in the new selection, or set to
the first line of the new selection if it is not.

Closes #1435.

git-svn-id: svn+ssh://aegisub/var/svn/aegisub/trunk@6398 93549f3f-7f0a-0410-a4b3-e966c9c94f04