Fix race condition that could result in a crash when opening video
When opening the video resulted in the script resolution changing, the
subtitles file was committed at a time when IsLoaded() would return true
but VideoOpen hadn't been signalled yet, resulting in some VideoDisplay
code running before the display was initialized. If the video opened is
sufficiently fast to decode, this could result in a crash due to trying
to display a frame before the display is shown.

git-svn-id: svn+ssh://aegisub/var/svn/aegisub/trunk@6645 93549f3f-7f0a-0410-a4b3-e966c9c94f04