Improve the negative-duration-avoiding behavior of the time edit boxes
Save the initial start and end times of lines before adjusting them to
avoid negative line durations, so that if further changes to the edit
boxes make the adjustment unneccesary the original values can be
restored. This Fixes the problem where changing a line visible from
frames 800-900 to 800-901 would actually result in 90-901, due to the
end frame briefly being 90.

git-svn-id: svn+ssh://aegisub/var/svn/aegisub/trunk@6741 93549f3f-7f0a-0410-a4b3-e966c9c94f04