Allow choosing the default style for new files.
This adds two new options, "Subtitle/Default Catalog" and
"Subtitle/Default Style", and a button, "Set as default", in the
Styles Manager.

New scripts will get the style "Subtitle/Default Style" if it is found in
the catalog "Subtitle/Default Catalog", otherwise the style "Default" if
it is found in the catalog "Default", otherwise the hardcoded style
"Default" (20px Arial with 2px shadow ugh).

Same for new styles created in the Styles Manager.