Check for audio hotkeys in both key down and char hook
wx doesn't call the keyboard hook when a window has capture, which
results in hotkeys not working while a mouse button is held down. This
usually doesn't matter much, but in the audio display one might want to
play audio while moving a marker. In general checking for hotkeys in
keydown doesn't work, but in the specific cast of marker dragging
keyboard focus is guaranteed to be on the audio display so it works out.

git-svn-id: svn+ssh://aegisub/var/svn/aegisub/trunk@6931 93549f3f-7f0a-0410-a4b3-e966c9c94f04