Mostly purge wxWidgets from non-UI stuff
Use boost::filesystem::path for all paths, and std::string for all other
strings, converting to/from wxString as close to the actual uses of wx
as possible.

Replace most of the uses of non-UI wxWidgets functionality with either
the equivalent libraries from boost, or (in a few cases), custom
replacements in libaegisub.

This touches pretty close to every file in Aegisub and is a pretty
radical change, so there's probably a lot of broken stuff.